refactor: move databases mod to tracker mod
Database is only use inisde the `tracker` mod.
