+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod databases;
 pub mod jobs;
 pub mod logging;
 pub mod protocol;
