+++ src/tracker/databases/error.rs
